test(#3590): Move integration tests to new lint test file
